requirements.txt
groq
pandas
pydantic
pydantic-settings
python-dotenv
tenacity
rich
gspread
google-auth
pytest